fix: se arregla el IVentasService

+++ src/main/java/com/joseAmbrocio/KinalApp/service/IVentasService.java
@@ -9,15 +9,15 @@ public interface IVentasService {
 
     List<Ventas> listarTodos();
 
-    Optional<Ventas> buscarVenta(int codigoVenta);
+    Optional<Ventas> buscarVenta(Long codigoVenta);
 
     Ventas guardarVenta(Ventas ventas);
 
-    Ventas actualizarVenta(int codigoVenta,Ventas ventas);
+    Ventas actualizarVenta(Long codigoVenta,Ventas ventas);
 
     void eliminarVenta(int codigoVenta);
 
-    boolean existeVenta(int codigoVenta);
+    boolean existeVenta(Long codigoVenta);
 
     List<Ventas> activos();
